Seperate mapService into own file
File cleanup
