+++ breach/src/map/map.h
@@ -4,9 +4,6 @@
 #include <vector>
 #include <string>
 
-#include "shade/game/entity/entity.h"
-#include "shade/game/world.h"
-#include "shade/instance/service/service.h"
 #include "shade/resource/handle.h"
 #include "shade/resource/resource.h"
 
@@ -33,25 +30,4 @@ private:
     std::string mName;
     std::vector<BackgroundElement> mBackgrounds;
     MapLayout mLayout;
-};
-
-// ======================================
-// TODO: Maybe separate into it's own file?
-class MapService : public Shade::Service {
-public:
-    MapService();
-    ~MapService();
-
-    // TODO: Make this somehow not have to pass in the entire game world to create the entities
-    //  - Perhaps an entity factory can be passed in for creating entities
-    //  - Alternatively, register entity factory as a service and locate the service to create new entities
-    std::vector<std::unique_ptr<Shade::Entity>> LoadMap(const std::string& path, Shade::GameWorldModule& gameWorld);
-
-    const MapLayout& GetLayout() const;
-
-protected:
-    void SetMapLayout(MapLayout layout);
-
-private:
-    MapLayout mCurrentLayout;
 };
